cleaned up module structure
