--- src/data/number.rs
@@ -1,7 +1,0 @@
-use serde::{Serialize, Deserialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Number(f64);
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct NumberDiff(f64);
